{{cookiecutter.repo_name}}/pyproject.toml
[tool.poetry]
name = "ml-cookiecutter-template"
version = "0.1.0"
description = ""
authors = ["Aziz Abdraimov <aziz.fizmat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
configargparse = "^1.7"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
